FLIT Utiva Project 1
This project, centered on market basket analysis, shows how certain items in a supermarket, when placed together, improve sales for each item. This project uses item frequency to show the number of times an item has been purchased, thereby making it possible for business/supermarket owners and retailers to group items together. However, this analysis is not only centered on products sales but also for those offering services, be it consultancy service, distribution service, car rentals etc. to create packages based on level of demand.